requirements.txt
pandas>=1.5
numpy>=1.23
PyYAML>=6.0
joblib>=1.2
scikit-learn>=1.2
matplotlib>=3.7
xarray>=2023.1.0
netCDF4>=1.6
rasterio>=1.3
